+++ CHANGELOG.md
@@ -10,6 +10,10 @@ is needed for these methods. Default: `false`, `true`, `true`, `true`.
 - New option `api:getkey` which allows you to set a different key to use in GET requests, if `privacy:get`. Default = `nil`.
 To disable this, set it back to nil: `$ midb set api:getkey nil`.
 - Exception handling to database operations - after testing, there's no way a request can break or stop the server. The only reason an exception will be risen is due to malformation in JSON files.
+- New GET verbs: `/api/endpoint/field` behaves like the normal `/api/endpoint/` but only returns the field specified in
+the JSON hash; `/api/endpoint/field/pattern` returns a JSON hash of all the entries where the field matches the pattern
+(a regular SQL `LIKE` pattern, by default it's `%pattern%`).
+- More verbose output on the server console; more info > less info!
 
 ### Changed
 - More compacted code in the server engine. Some complicated logics were moved to methods.